refactor(Game): change amount data type from Integer to Long
Migrated all monetary value representations from Integer to Long to
accommodate larger sums and to prevent potential overflows. This change
affects the purchase amount, lotto ticket costs, and the calculations of
winnings and return on investment.

+++ src/main/java/lotto/constants/Constants.java
@@ -1,7 +1,7 @@
 package lotto.constants;
 
 public class Constants {
-    public static final Integer LOTTO_PRICE = 1000;
+    public static final Long LOTTO_PRICE = 1000L;
     public static final Integer START = 1;
     public static final Integer END = 45;
     public static final Integer NUMBER_COUNT = 6;
